2015-03-23 18:56 UTC+0100 Viktor Szakats (vszakats users.noreply.github.com)   * contrib/gtwvw/gtwvwd.c   * contrib/gtwvw/tooltip.c   * contrib/gtwvw/wvwdraw.c   * contrib/hbmemio/memio.c   * contrib/hbmzip/3rd/minizip/minizip.dif   * contrib/hbmzip/3rd/minizip/unzip.c     ! fix problems detected by msvc
  * contrib/hbpost.hbm
    ! honor undocumented option to disable '-analyze' option for msvc

+++ contrib/gtwvw/tooltip.c
@@ -178,7 +178,7 @@ HB_FUNC( WVW_SETTOOLTIPWIDTH )
 
    if( wvw_win )
    {
-      hb_retni( SendMessage( wvw_win->hWndTT, TTM_GETMAXTIPWIDTH, 0, 0 ) );
+      hb_retni( ( int ) SendMessage( wvw_win->hWndTT, TTM_GETMAXTIPWIDTH, 0, 0 ) );
 
       if( HB_ISNUM( 2 ) )
          SendMessage( wvw_win->hWndTT, TTM_SETMAXTIPWIDTH, 0, ( LPARAM ) hb_parni( 2 ) );
@@ -251,7 +251,7 @@ HB_FUNC( WVW_GETTOOLTIPWIDTH )
 
    if( wvw_win )
    {
-      hb_retni( SendMessage( wvw_win->hWndTT, TTM_GETMAXTIPWIDTH, 0, 0 ) );
+      hb_retni( ( int ) SendMessage( wvw_win->hWndTT, TTM_GETMAXTIPWIDTH, 0, 0 ) );
       return;
    }
 #endif
